pyproject.toml
[project]
name = "ookla-speed-test-from-s3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.28.0",
    "delta-spark==3.2.0",
    "deltalake>=1.2.1",
    "minio==7.1.16",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyspark==3.5.1",
    "pyyaml>=6.0.1",
]
